+++ src/@tests/functional/zanix.test.ts
@@ -1,8 +1,8 @@
 import { getAllZanixLibrariesInfo, getLatestRelease } from '../../modules/helpers/zanix/info.ts'
 import { versionRegex } from 'utils/regex.ts'
 import { assertMatch } from '@std/assert/assert-match'
-//import { getZanixPaths } from 'modules/helpers/zanix/tree.ts'
-//import { assert } from '@std/assert'
+import { getZanixPaths } from 'modules/helpers/zanix/tree.ts'
+import { assert } from '@std/assert'
 
 Deno.test('Fetching Zanix lates version validation', async () => {
   const result = await getLatestRelease('utils')
@@ -21,16 +21,28 @@ Deno.test('Fetching Zanix lates release validation', async () => {
   assertMatch(result['@zanix/server'].version, versionRegex)
   assertMatch(result['@zanix/tasker'].version, versionRegex)
 })
-/*
-// TODO: habilite when remove alpha
-Deno.test('getZanixPaths should return correct constants content from jsr', async () => {
+
+Deno.test('getZanixPaths should return correct default content from jsr', async () => {
   const paths = getZanixPaths('library', '')
 
-  paths.subfolders.src.subfolders.utils.templates.base[0].PATH = '/src/utils/constants.ts'
-  const _content = await paths.subfolders.src.subfolders.utils.templates.base[0].content({
+  const contentUtils = await paths.subfolders.src.subfolders.utils.templates.base[0].content({
     metaUrl: import.meta.url,
   })
-  // TODO: habilite when remove alpha
-  assert(content !== '')
-  //assert(content.includes('ZANIX_LOGO'))
-})*/
+
+  assert(contentUtils !== '')
+  assert(contentUtils.includes('Utilities Module Template'))
+
+  const contentMod = await paths.templates.base[1].content({
+    metaUrl: import.meta.url,
+  })
+
+  assert(contentMod !== '')
+  assert(contentMod.includes('Module Template'))
+
+  const contentLicense = await paths.subfolders.docs.templates.base[1].content({
+    metaUrl: import.meta.url,
+  })
+
+  assert(contentLicense !== '')
+  assert(contentLicense.includes('License'))
+})